Store snap data in tool data
