+++ editor/src/tool/tools/pen.rs
@@ -1,4 +1,5 @@
 use crate::input::InputPreprocessor;
+use crate::tool::snapping::SnapHandler;
 use crate::tool::{DocumentToolData, Fsm, ToolActionHandlerData, ToolOptions, ToolType};
 use crate::{document::DocumentMessageHandler, message_prelude::*};
 use glam::DAffine2;
@@ -52,6 +53,7 @@ struct PenToolData {
 	next_point: DAffine2,
 	weight: u32,
 	path: Option<Vec<LayerId>>,
+	snap_handler: SnapHandler,
 }
 
 impl Fsm for PenToolFsmState {
@@ -60,7 +62,7 @@ impl Fsm for PenToolFsmState {
 	fn transition(
 		self,
 		event: ToolMessage,
-		document: &mut DocumentMessageHandler,
+		document: &DocumentMessageHandler,
 		tool_data: &DocumentToolData,
 		data: &mut Self::ToolData,
 		input: &InputPreprocessor,
@@ -77,8 +79,8 @@ impl Fsm for PenToolFsmState {
 					responses.push_back(DocumentMessage::DeselectAllLayers.into());
 					data.path = Some(vec![generate_uuid()]);
 
-					document.snapping_handler.start_snap(&document.graphene_document, document.all_layers_sorted(), &[]);
-					let snapped_position = document.snapping_handler.snap_position(input.mouse.position);
+					data.snap_handler.start_snap(document, document.all_layers_sorted(), &[]);
+					let snapped_position = data.snap_handler.snap_position(document, input.mouse.position);
 
 					let pos = transform.inverse() * DAffine2::from_translation(snapped_position);
 
@@ -93,7 +95,7 @@ impl Fsm for PenToolFsmState {
 					Dragging
 				}
 				(Dragging, DragStop) => {
-					let snapped_position = document.snapping_handler.snap_position(input.mouse.position);
+					let snapped_position = data.snap_handler.snap_position(document, input.mouse.position);
 					let pos = transform.inverse() * DAffine2::from_translation(snapped_position);
 
 					// TODO: introduce comparison threshold when operating with canvas coordinates (https://github.com/GraphiteEditor/Graphite/issues/100)
@@ -107,7 +109,7 @@ impl Fsm for PenToolFsmState {
 					Dragging
 				}
 				(Dragging, PointerMove) => {
-					let snapped_position = document.snapping_handler.snap_position(input.mouse.position);
+					let snapped_position = data.snap_handler.snap_position(document, input.mouse.position);
 					let pos = transform.inverse() * DAffine2::from_translation(snapped_position);
 					data.next_point = pos;
 
@@ -126,6 +128,7 @@ impl Fsm for PenToolFsmState {
 
 					data.path = None;
 					data.points.clear();
+					data.snap_handler.cleanup();
 
 					Ready
 				}
@@ -133,6 +136,7 @@ impl Fsm for PenToolFsmState {
 					responses.push_back(DocumentMessage::AbortTransaction.into());
 					data.points.clear();
 					data.path = None;
+					data.snap_handler.cleanup();
 
 					Ready
 				}
