+++ editor/src/tool/tools/shape.rs
@@ -59,7 +59,7 @@ impl Fsm for ShapeToolFsmState {
 	fn transition(
 		self,
 		event: ToolMessage,
-		document: &mut DocumentMessageHandler,
+		document: &DocumentMessageHandler,
 		tool_data: &DocumentToolData,
 		data: &mut Self::ToolData,
 		input: &InputPreprocessor,
@@ -109,12 +109,12 @@ impl Fsm for ShapeToolFsmState {
 						false => responses.push_back(DocumentMessage::CommitTransaction.into()),
 					}
 
-					shape_data.path = None;
+					shape_data.cleanup();
 					Ready
 				}
 				(Dragging, Abort) => {
 					responses.push_back(DocumentMessage::AbortTransaction.into());
-					shape_data.path = None;
+					shape_data.cleanup();
 
 					Ready
 				}
